- rewriting api V2

+++ cocroach_utils/db_conv.py
@@ -1,51 +1,39 @@
 # This file contains all the functions for the conversations table
 from cocroach_utils.db_history import delete_history_by_conv_id
-from cocroach_utils.database_utils import connect_to_db, save_error
+from cocroach_utils.database_utils import connect_to_db, save_error, get_db_cursor, fetch_all, fetch_one
 from cocroach_utils.db_users import get_user_by_id
 
 
-def get_all_conversations(user_id):
+def get_all_docs():
     """
-    Get all conversations for admin
+    Get all docs
     :return:
     """
-    # check if this user has access to this conversation
+    with get_db_cursor() as cursor:
+        if cursor:
+            return fetch_all(cursor, "SELECT * FROM documents")
+    return []
+
+
+def get_all_conversations_by_user(user_id):
+    """
+    Get all conversations for admin
+    :param user_id:
+    :return:
+    """
     user = get_user_by_id(user_id)
     if user is None:
         if user["role"] != 10:
             return []
 
-    conn = connect_to_db()
-    if conn is not None:
-        try:
-            with conn.cursor() as cur:
-                # add to below also assistant.title from assistants table by conversations.assistant
-                cur.execute(
-                    "SELECT conv_id, user_id, doc_id, conversations.title, active, summary, models.name, "
-                    "assistants.title FROM conversations LEFT JOIN models ON conversations.model = models._id LEFT "
-                    "JOIN assistants ON conversations.assistant = assistants._id",
-                    (user_id,))
-                conn.commit()
-                docs = []
-                for doc in cur.fetchall():
-                    docs.append({
-                        "conv_id": str(doc[0]),
-                        "user_id": str(doc[1]),
-                        "doc_id": str(doc[2]),
-                        "title": doc[3],
-                        "active": doc[4],
-                        "summary": doc[5],
-                        "model": doc[6],
-                        "assistant": doc[7]
-                    })
-                return docs
-        except Exception as err:
-            conn.rollback()
-            save_error(err)
-            return []
-    else:
-        save_error("No connection to the database")
-        return []
+    with get_db_cursor() as cursor:
+        if cursor:
+            return fetch_all(cursor,
+                             "SELECT conv_id, user_id, doc_id, conversations.title, active, summary, models.name, "
+                             "assistants.title, memory FROM conversations "
+                             "LEFT JOIN models ON conversations.model = models._id LEFT "
+                             "JOIN assistants ON conversations.assistant = assistants._id", (user_id,))
+    return []
 
 
 def get_user_conversations(user_id):
@@ -54,38 +42,15 @@ def get_user_conversations(user_id):
     :param user_id:
     :return:
     """
-    conn = connect_to_db()
-    if conn is not None:
-        try:
-            with conn.cursor() as cur:
-                cur.execute(
-                    "SELECT conv_id, user_id, doc_id, conversations.title, active, summary, models.name, "
-                    "assistants.title, assistants.system_prompt FROM conversations LEFT JOIN models ON conversations.model = models._id LEFT "
-                    "JOIN assistants ON conversations.assistant = assistants._id "
-                    "WHERE conversations.user_id = %s",
-                    (user_id,))
-                conn.commit()
-                docs = []
-                for doc in cur.fetchall():
-                    docs.append({
-                        "conv_id": str(doc[0]),
-                        "user_id": str(doc[1]),
-                        "doc_id": str(doc[2]),
-                        "title": doc[3],
-                        "active": doc[4],
-                        "summary": doc[5],
-                        "model": doc[6],
-                        "assistant": doc[7],
-                        "system_prompt": doc[8]
-                    })
-                return docs
-        except Exception as err:
-            conn.rollback()
-            save_error(err)
-            return []
-    else:
-        save_error("No connection to the database")
-        return []
+    with get_db_cursor() as cursor:
+        if cursor:
+            return fetch_all(cursor,
+                             "SELECT conv_id, user_id, doc_id, conversations.title, active, summary, models.name, "
+                             "assistants.title, assistants.system_prompt, memory FROM conversations "
+                             "LEFT JOIN models ON conversations.model = models._id LEFT "
+                             "JOIN assistants ON conversations.assistant = assistants._id "
+                             "WHERE conversations.user_id = %s", (user_id,))
+    return []
 
 
 def get_conv_by_id(conversation_id):
@@ -94,36 +59,15 @@ def get_conv_by_id(conversation_id):
     :param conversation_id:
     :return:
     """
-    conn = connect_to_db()
-    if conn is not None:
-        try:
-            with conn.cursor() as cur:
-                cur.execute(
-                    "SELECT conv_id, user_id, doc_id, conversations.title, active, summary, models.name, "
-                    "assistants.title, assistants.system_prompt FROM conversations LEFT JOIN models ON conversations.model = models._id LEFT "
-                    "JOIN assistants ON conversations.assistant = assistants._id "
-                    "WHERE conv_id = %s",
-                    (conversation_id,))
-                conn.commit()
-                doc = cur.fetchone()
-                return {
-                    "conv_id": str(doc[0]),
-                    "user_id": str(doc[1]),
-                    "doc_id": str(doc[2]),
-                    "title": doc[3],
-                    "active": doc[4],
-                    "summary": doc[5],
-                    "model": doc[6],
-                    "assistant": doc[7],
-                    "system_prompt": doc[8]
-                }
-        except Exception as err:
-            conn.rollback()
-            save_error(err)
-            return []
-    else:
-        save_error("No connection to the database")
-        return []
+    with get_db_cursor() as cursor:
+        if cursor:
+            return fetch_one(cursor,
+                             "SELECT conv_id, user_id, doc_id, conversations.title, active, summary, models.name, "
+                             "assistants.title, assistants.system_prompt, memory FROM conversations "
+                             "LEFT JOIN models ON conversations.model = models._id LEFT "
+                             "JOIN assistants ON conversations.assistant = assistants._id "
+                             "WHERE conv_id = %s", (conversation_id,))
+    return []
 
 
 def add_conversation(user_id, doc_id=None, title="New conversation", model=0, assistant=0):
@@ -136,22 +80,12 @@ def add_conversation(user_id, doc_id=None, title="New conversation", model=0, as
     :param title:
     :return: conv_id
     """
-    conn = connect_to_db()
-    if conn is not None:
-        try:
-            with conn.cursor() as cur:
-                cur.execute(
-                    "INSERT INTO conversations (user_id, doc_id, title, model, assistant) VALUES (%s, %s, %s, %s, %s) RETURNING conv_id",
-                    (user_id, doc_id, title, model, assistant))
-                conn.commit()
-                return str(cur.fetchone()[0])
-        except Exception as err:
-            conn.rollback()
-            save_error(err)
-            return -1
-    else:
-        save_error("No connection to the database")
-        return -1
+    with get_db_cursor() as cursor:
+        if cursor:
+            return fetch_one(cursor,
+                             "INSERT INTO conversations (user_id, doc_id, title, model, assistant) VALUES (%s, %s, %s, "
+                             "%s, %s) RETURNING conv_id", (user_id, doc_id, title, model, assistant))['conv_id']
+    return -1
 
 
 def update_conversation_field(conversation_id, field, value):
@@ -162,22 +96,12 @@ def update_conversation_field(conversation_id, field, value):
     :param value:
     :return:
     """
-    conn = connect_to_db()
-    if conn is not None:
-        try:
-            with conn.cursor() as cur:
-                cur.execute(
-                    "UPDATE conversations SET " + field + " = %s WHERE conv_id = %s",
-                    (value, conversation_id))
-                conn.commit()
-                return True
-        except Exception as err:
-            conn.rollback()
-            save_error(err)
-            return False
-    else:
-        save_error("No connection to the database")
-        return False
+    with get_db_cursor() as cursor:
+        if cursor:
+            cursor.execute("UPDATE conversations SET " + field + " = %s WHERE conv_id = %s",
+                           (value, conversation_id))
+            return cursor.rowcount == 1
+    return False
 
 
 def delete_conversation(conversation_id):
@@ -186,23 +110,11 @@ def delete_conversation(conversation_id):
     :param conversation_id:
     :return:
     """
-    conn = connect_to_db()
-    if conn is not None:
-        try:
-            delete_history_by_conv_id(conversation_id)
-            with conn.cursor() as cur:
-                cur.execute(
-                    "DELETE FROM conversations WHERE conv_id = %s",
-                    (conversation_id,))
-                conn.commit()
-                return True
-        except Exception as err:
-            conn.rollback()
-            save_error(err)
-            return False
-    else:
-        save_error("No connection to the database")
-        return False
+    with get_db_cursor() as cursor:
+        if cursor:
+            cursor.execute("DELETE FROM conversations WHERE conv_id = %s", (conversation_id,))
+            return cursor.rowcount == 1
+    return False
 
 
 def delete_conversation_by_user(user_id, delete_history=False):
@@ -219,19 +131,9 @@ def delete_conversation_by_user(user_id, delete_history=False):
         for conv in conversations:
             delete_history_by_conv_id(conv[0])
 
-    conn = connect_to_db()
-    if conn is not None:
-        try:
-            with conn.cursor() as cur:
-                cur.execute(
-                    "DELETE FROM conversations WHERE user_id = %s",
-                    (user_id,))
-                conn.commit()
-                return True
-        except Exception as err:
-            conn.rollback()
-            save_error(err)
-            return False
-    else:
-        save_error("No connection to the database")
-        return False
+    with get_db_cursor() as cursor:
+        if cursor:
+            cursor.execute("DELETE FROM conversations WHERE user_id = %s", (user_id,))
+            return cursor.rowcount == 1
+    return False
+
